Create prometheus stats dir outside OMERO.web dir
This ensures prometheus_multiproc_dir is not lost on upgrade, which may lead to OMERO.web failing to restart since some ExecStartPre bin/omero steps may require the presence of the dir even if web isn't yet running.

This also deletes the contents of prometheus_multiproc_dir isntead of deleting and recreating the dir itself.